added signals to update a game rating with relative tests

+++ fiordispino/models/game.py
@@ -23,7 +23,7 @@ class Game(models.Model):
     box_art = models.ImageField(upload_to=build_path)
     description = models.TextField(validators=[validate_game_description])
     title = models.TextField(validators=[validate_title])
-    global_rating = models.FloatField(default=0.0, validators=[validate_global_rating])
+    global_rating = models.DecimalField(default=0.0, max_digits=3, decimal_places=1, validators=[validate_global_rating])
 
     # to count how many players reviewed it. No need for custom validation, it's implicit thanks to PositiveInteger
     rating_count = models.PositiveIntegerField(default=0)
